+++ config/jest/jestMockUps.js
@@ -1,0 +1,9 @@
+jest.mock('react-i18next', () => ({
+    // this mock makes sure any components using the translate HoC receive the t function as a prop
+    withTranslation: () => Component => {
+        Component.defaultProps = { ...Component.defaultProps, t: () => "" };
+        return Component;
+    },
+}));
+
+jest.mock('../../src/i18n');