+++ src/layouts/Dashboard/Dashboard.spec.tsx
@@ -46,17 +46,13 @@ describe('Dashboard layout',()=>{
     let shallow;
     let mount;
 
-    beforeAll(() => {
-        mount = createMount();
-    });
-
-    afterAll(() => {
-        mount.cleanUp();
-    });
-
-
     beforeEach(() => {
+        mount = createMount();
         shallow = createShallow();
+    });
+
+    afterEach(()=>{
+        mount.cleanUp();
     });
 
     it('render in normal state correctly by default',()=>{